⚡️ 不必要のデータ取得を削除、またarticlesファイルの名前変更
